requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==10.4.0
potracer==0.0.4
